test(RegisterTenantForm):
