+++ app/Observers/TenantObserver.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Observers;
+
+use App\Tenant;
+
+class TenantObserver
+{
+    public function created(Tenant $tenant): void
+    {
+    }
+}
