--- resources/views/livewire/pages/auth/confirm-password.blade.php
@@ -1,65 +1,0 @@
-<?php
-
-use App\Providers\RouteServiceProvider;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Validation\ValidationException;
-
-use function Livewire\Volt\layout;
-use function Livewire\Volt\rules;
-use function Livewire\Volt\state;
-
-layout('layouts.guest');
-
-state(['password' => '']);
-
-rules(['password' => ['required', 'string']]);
-
-$confirmPassword = function () {
-    $this->validate();
-
-    if (! Auth::guard('web')->validate([
-        'email' => Auth::user()->email,
-        'password' => $this->password,
-    ])) {
-        throw ValidationException::withMessages([
-            'password' => __('auth.password'),
-        ]);
-    }
-
-    session(['auth.password_confirmed_at' => time()]);
-
-    $this->redirect(
-        session('url.intended', RouteServiceProvider::HOME),
-        navigate: true
-    );
-};
-
-?>
-
-<div>
-    <div class="mb-4 text-sm text-gray-600 dark:text-gray-400">
-        {{ __('This is a secure area of the application. Please confirm your password before continuing.') }}
-    </div>
-
-    <form wire:submit="confirmPassword">
-        <!-- Password -->
-        <div>
-            <x-input-label for="password" :value="__('Password')" />
-
-            <x-text-input wire:model="password"
-                          id="password"
-                          class="block mt-1 w-full"
-                          type="password"
-                          name="password"
-                          required autocomplete="current-password" />
-
-            <x-input-error :messages="$errors->get('password')" class="mt-2" />
-        </div>
-
-        <div class="flex justify-end mt-4">
-            <x-primary-button>
-                {{ __('Confirm') }}
-            </x-primary-button>
-        </div>
-    </form>
-</div>
