+++ routes/web.php
@@ -1,10 +1,13 @@
 <?php
 
+use App\Livewire\Forms\RegisterTenantForm;
 use Illuminate\Support\Facades\Route;
 use Livewire\Volt\Volt;
 
 Volt::route('/', 'pages/welcome');
-Volt::route('/auth/register', 'pages/auth/register')->name('auth.register');
+
+Route::get('/auth/register', RegisterTenantForm::class)
+    ->name('auth.register');
 
 Route::middleware([
     'web',
